Organizing edit-dialog.component structure
