+++ employee-manager/src/app/components/employees-view/employees-view.component.css
@@ -127,7 +127,7 @@ tr:hover {
 .progress-bar-value {
     width: 100%;
     height: 100%;
-    background: linear-gradient(90deg, #3893f4 0%, #4ab160 100%);
+    background: linear-gradient(90deg, #3893f4 0%, #3893f4 100%);
     animation: indeterminate 1.5s infinite linear;
     transform-origin: 0% 50%;
 }
